Fixed precision warnings in Gumbo library

+++ ObjectiveGumbo/Gumbo/error.c
@@ -33,7 +33,7 @@
 static int print_message(
     GumboParser* parser, GumboStringBuffer* output, const char* format, ...) {
   va_list args;
-  int remaining_capacity = output->capacity - output->length;
+  int remaining_capacity = (int)(output->capacity - output->length);
   va_start(args, format);
   int bytes_written = vsnprintf(
       output->data + output->length, remaining_capacity, format, args);
@@ -62,7 +62,7 @@ static int print_message(
   if (bytes_written > remaining_capacity) {
     gumbo_string_buffer_reserve(
         parser, output->capacity + bytes_written, output);
-    remaining_capacity = output->capacity - output->length;
+    remaining_capacity = (int) (output->capacity - output->length);
     va_start(args, format);
     bytes_written = vsnprintf(
         output->data + output->length, remaining_capacity, format, args);
